Add new module for BOMs in other projects
The top level project here now is spring-cloud-build and any library
in Spring Cloud should try and use it as a parent to avoid duplicate
plugin and build declarations.

The new module here is spring-cloud-dependencies-parent and all
BOMs in Spring Cloud should use this as a parent (for the same reason).
All sub-projects in Spring Cloud should have a BOM, and if they want to
(usually they do) they can use it to define dependencies by importing
it into their parent.

User's apps should use the spring-cloud-*-dependencies if they need
a BOM (because they do *not* have spring-boot-* as a parent). Then
they can choose a version of Spring Boot more easily (by choosing
a version of spring-boot-starter-parent).