Add spring-cloud-build-dependencies BOM layer
Users, and even Spring Cloud libraries never need to know about this,
but the extra level of indirection helps because of the commons-logging
dependency, which we want to exclude from Spring Cloud libraries, but
which can only be excluded globally from all of them here. The previous
solution to that problem was to add the exclusion with a spring-core
dependency managed by spring-cloud-build directly. The problem with that
is that it needs to have the spring-core version explicitly, or else
libraries that depend directly on spring-core cannot find the version
(not many libraries *do* depend directly on spring-core so it only hurts
on the margin to not have the version, which is where we were with
1.3.1.RELEASE and lower).